requirements.txt
factorio-rcon-py
